Implement the functionality for todo list.
Added the form handling functionality such as submitting input,
cancelling input, and clearing input. Once the input is submitted, it will be
stored in an array, and be stored in object for organization.
From the array, a display card functionality is implemented which
fetch data from the array and put it in a card to be displayed in the page.