Add functionalities for the project
Create project page when the project button in nav is pressed. Create
add project button, change page name, and show the project modal when the
add project button is pressed. Store the data from project form, and use it
to display a card. Store the data in the localStorage, and load data
from localStorage when the project page is loaded.