Add file to handle a project card todo contents
Add the functionalities to handle creation of todo page for each project cards.
This will also get the project name and list the todo card in a new array.
The array will be stored in a localStorage, where the array elements can be saved and loaded.
Also added functionality to handle each todo card. Every todo card that has the
current project name as their assigned project will be displayed and put into the array.