Add functionality for edit button
If in a To-do page, both in general and each project,
the edit button will show an edit modal for the task.
If in a project page, the edit button will show the
edit modal for project. Once confirm, the localStorage and
arrays that store the data will be updated, and display the
updated data on the card.