Remove storage handling for each project
The storage handler for each project is redundant, just use the
storage handler in the todo. Also, resolve the currentProjectName
persistence issue, which cause a bug where when the user click the
Todo Button (in the nav buttons) coming from a project todo card view,
the project name persist as the context (assignedProject) instead of
reverting to "General" which is the intended behavior. By adding a
reset function for currentProjectName, this issue is resolved.