Add delete and edit handler, edit form handler
The delete and edit btn handler is placed here in dom_manipulation for
better code organization. The dialog classes is being toggled to add or removed
class="hidden" for explicit control of showing or closing modals. Add preventDefault
and stopPropagation to form cancel btn handler to not show the error message when
the user cancel the form.